remove dependencies in favor of code duplication
