+++ gateway-js/src/index.ts
@@ -1271,4 +1271,3 @@ export {
 };
 
 export * from './datasources';
-export * from './schema-helper';
