+++ subgraph-js/src/schema-helper/index.ts
@@ -1,3 +1,5 @@
 export * from './resolverMap';
 export * from './buildSchemaFromSDL';
 export * from './error';
+export * from './transformSchema';
+
