+++ subgraph-js/src/schema-helper/transformSchema.ts
