+++ subgraph-js/tsconfig.json
@@ -6,7 +6,5 @@
   },
   "include": ["src/**/*"],
   "exclude": ["**/__tests__"],
-  "references": [
-    { "path": "../gateway-js" },
-  ]
+  "references": []
 }
